backend/requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==3.0.4
django-cors-headers==3.2.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
httpie==2.0.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.2
pandas==1.0.3
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
